refactor: move conf_parser out of alert_utils
